Revert "Example tunning"
