+++ examples/blog/Makefile
@@ -1,8 +1,6 @@
 .PHONY: build
 
 install:
-	@mkdir -p build
-	@cp ../../build/* ./build
 	npm install
 	bower install
 
@@ -11,6 +9,4 @@ build:
 	@echo "File build/app.js updated"
 
 run:
-	./node_modules/webpack-dev-server/bin/webpack-dev-server.js \
-		--progress --colors --hot \
-		--content-base .
+	./node_modules/webpack-dev-server/bin/webpack-dev-server.js --port 8088 --progress --colors --hot --content-base .
