+++ examples/blog/bower.json
@@ -1,0 +1,20 @@
+{
+  "name": "react-admin",
+  "version": "0.0.1",
+  "authors": [
+    "marmelab <info@marmelab.com>"
+  ],
+  "description": "React-admin blog backend demo",
+  "license": "MIT",
+  "private": true,
+  "ignore": [
+    "**/.*",
+    "bower_components"
+  ],
+  "dependencies": {
+    "fakerest": "master",
+    "react": "~0.13.*",
+    "sinon-server": "http://sinonjs.org/releases/sinon-server-1.14.1.js",
+    "react-router": "~0.13.3"
+  }
+}
