--- examples/blog/js/utils.js
@@ -1,9 +1,0 @@
-
-
-export function truncate(value) {
-    if (!value) {
-        return '';
-    }
-
-    return value.length > 50 ? value.substr(0, 50) + '...' : value;
-}
