+++ examples/blog/test.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>React-Admin Blog Sample Application</title>
+    <link rel="stylesheet" href="build/react-admin-standalone.min.css">
+</head>
+<body>
+    <div id="my-app"></div>
+
+    <script src="build/babel.min.js" type="text/javascript"></script>
+    <script src="bower_components/react/react.js" type="text/javascript"></script>
+    <script src="bower_components/react-router/build/umd/ReactRouter.min.js" type="text/javascript"></script>
+    <script src="bower_components/fakerest/dist/FakeRest.min.js" type="text/javascript"></script>
+    <script src="bower_components/sinon-server/index.js" type="text/javascript"></script>
+    <script src="data.js" type="text/javascript"></script>
+    <script src="fakerest-init.js" type="text/javascript"></script>
+
+    <script src="build/react-admin-standalone.min.js"></script>
+    <script src="config.js" type="text/babel"></script>
+</body>
+</html>
