+++ examples/blog/webpack.config.js
@@ -1,22 +1,29 @@
-var path = require('path');
+function getEntrySources() {
+    var sources = [];
+
+    if (process.env.NODE_ENV !== 'production') { // for live reload
+        sources.push('webpack-dev-server/client?http://0.0.0.0:8088');
+        sources.push('webpack/hot/dev-server');
+    }
+    sources.push('./config-webpack.js'); // must be the last one
+
+    return sources;
+}
 
 module.exports = {
-  entry: {
-    main: './js/main.js',
-  },
-  output: {
-    path: __dirname + '/build',
-    filename: '[name].js'
-  },
-  module: {
-    loaders: [{
-      test: /\.jsx?$/,
-      loader: 'babel',
-      include: [
-        path.resolve(__dirname, "js")
-      ]
-    }]
-  },
-  plugins: [
-  ]
+    entry: {
+        'app': getEntrySources()
+    },
+    output: {
+        path: __dirname + '/build',
+        filename: '[name].js'
+    },
+    module: {
+        loaders: [
+            { test: /config-webpack\.js$/, loaders: ['react-hot', 'babel'] }
+        ]
+    },
+    node: {
+        fs: 'empty'
+    }
 };
